refactor: change Price and OriginalPrice types from int to decimal for accurate financial calculations

+++ frontend/Models/Api/Event.cs
@@ -6,8 +6,8 @@ namespace GloboTicket.Frontend.Models.Api
     {
         public Guid EventId { get; set; }
         public string Name { get; set; }
-        public int Price { get; set; }
-        public int OriginalPrice { get; set; }
+        public decimal Price { get; set; }
+        public decimal OriginalPrice { get; set; }
         public string Artist { get; set; }
         public DateTime Date { get; set; }
         public string Description { get; set; }
