+++ frontend/Models/View/BasketLineViewModel.cs
@@ -8,8 +8,8 @@ namespace GloboTicket.Frontend.Models.View
         public Guid EventId { get; set; }
         public string EventName { get; set; }
         public DateTimeOffset Date { get; set; }
-        public int Price { get; set; }
-        public int OriginalPrice { get; set; }
+        public decimal Price { get; set; }
+        public decimal OriginalPrice { get; set; }
         public bool IsOnSpecialOffer { get; set; }
         public int Quantity { get; set; }
     }
